refactor: [FatSecretApiClient] 패키지 구조 변경
