--- src/main/java/com/hotnerds/fatsecret/exception/FatSecretResponseError.java
@@ -1,14 +1,0 @@
-package com.hotnerds.fatsecret.exception;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonRootName;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-public class FatSecretResponseError {
-    private Integer code;
-    private String message;
-}
