+++ src/main/java/com/hotnerds/food/infrastructure/FatSecretApiClient.java
@@ -1,28 +1,19 @@
-package com.hotnerds.fatsecret.application;
+package com.hotnerds.food.infrastructure;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.hotnerds.fatsecret.FatSecretToken;
+import com.hotnerds.food.infrastructure.FatSecretToken;
 
 import java.net.URI;
 import java.util.Map;
 
-import com.hotnerds.fatsecret.exception.FatSecretResponseErrorHandler;
-import lombok.Getter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.*;
-import org.springframework.http.client.BufferingClientHttpRequestFactory;
-import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.stereotype.Component;
-import org.springframework.web.client.ResourceAccessException;
 import org.springframework.web.client.RestClientResponseException;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
-
-import javax.annotation.Resource;
-import javax.print.attribute.standard.Media;
 
 @Component
 public class FatSecretApiClient {
