+++ src/main/java/com/hotnerds/food/infrastructure/exception/FatSecretResponseError.java
@@ -1,0 +1,9 @@
+package com.hotnerds.food.infrastructure.exception;
+
+import lombok.Getter;
+
+@Getter
+public class FatSecretResponseError {
+    private Integer code;
+    private String message;
+}
