+++ src/test/java/com/hotnerds/food/infrastructure/TestConfig.java
@@ -1,7 +1,6 @@
-package com.hotnerds.fatsecret;
+package com.hotnerds.food.infrastructure;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.hotnerds.fatsecret.exception.FatSecretResponseErrorHandler;
-import org.mockito.Mock;
+import com.hotnerds.food.infrastructure.exception.FatSecretResponseErrorHandler;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.test.web.client.MockServerRestTemplateCustomizer;
 import org.springframework.boot.web.client.RestTemplateBuilder;
@@ -9,7 +8,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.DependsOn;
 import org.springframework.http.client.BufferingClientHttpRequestFactory;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
-import org.springframework.web.client.ResponseErrorHandler;
 
 import static org.mockito.Mockito.mock;
 @TestConfiguration
